requirements.txt
Django>=4.2,<6.0
djangorestframework>=3.14.0
python-dotenv>=1.0.1
midtransclient>=1.4.2
psycopg2-binary>=2.9.9
dj-database-url>=2.1.0
